Change how we flatten arrays to appease babel
Previous to this change, it was transpiling to [...value]
which doesn't really do anything. It is a bit hard for us to
upgrade our dependencies (or at least it fought me a lot today)
so I think maybe we should just do it this way and move on with our
lives.