bug 1328719 - sort $json output.
Since we need to be able to deterministically rebuild task definitions
via json-e, we need the output to be deterministic. This means sorting
$json.

I'm pretty confident in the python fix, but would like extra attention
on the js fix. Go `json.Marshal()` appears to have its own sorting
rules (per the struct it's sorting? [1] [2]); I'm not sure if it's going
to match js and python.

[1] https://stackoverflow.com/questions/18668652/how-to-produce-json-with-sorted-keys-in-go
[2] https://golang.org/pkg/encoding/json/#Marshal