WIP on converting to serde_json
Serde's Value::Number seems intent on distinguishing integers from
floats, but does not make that distinction available in the API, so
there's no good wya to know that "3*7" should be "21" and not "21.0".

It may be that this is different with some (hidden) features enabled.
But this library should probably replicate the relevant features or risk
having incompatible types.